+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `statuses` getter
 - `setOnInsert` parameter in `save` and `multiSave` methods
+- `mapIdByReferenceId` method
 
 ## [3.5.0] - 2020-01-13
 ### Added